BlueSense/requirements.txt
streamlit==1.28.0
atproto==0.0.40
google-cloud-language==2.13.1
matplotlib==3.7.2
pandas==2.1.0
python-dotenv==1.0.0
requests==2.31.0
wordcloud==1.9.2 